Verified that SPI at 4MHz or higher can not be handled by interrupt due to too long interrupt response times. Current implementation has sticking interrupt of 302us per 2ms of sent SPI message. To do: - continue application implementation

+++ TrackController4.X/main.h
@@ -37,8 +37,8 @@ typedef struct
     unsigned int        HoldingRegWrSl[4];
     unsigned int        InputReg[2];
     unsigned int        DiagReg[4];
-    unsigned int        MbReceiveCounter;
-    unsigned int        MbSentCounter;
+    unsigned long int   MbReceiveCounter;
+    unsigned long int   MbSentCounter;
     SLAVE_DATA          MbCommError;
     unsigned char       MbExceptionCode;
     unsigned int        SpiCommErrorCounter;
